refactor: rename classes and methods for consistency, add validation to BackupController, and implement getLatestBackup endpoint
