+++ src/main/java/com/codeit/hrbank/domain/backup/sevice/BackupService.java
@@ -7,10 +7,8 @@ import jakarta.servlet.http.HttpServletRequest;
 
 public interface BackupService {
 
-CursorPageResponseBackupDto getBackupList(CursorBackupRequestDto cursorBackupRequestDto);
-BackupDto createBackup(HttpServletRequest request);
-BackupDto getLatestBackup();
-
-
+    CursorPageResponseBackupDto getBackupList(CursorBackupRequestDto cursorBackupRequestDto);
+    BackupDto createBackup(HttpServletRequest request);
+    BackupDto getLatestBackup();
 
 }
